requirements.txt
numpy
pyserial
pillow
scipy
pysinewave
opencv-python
